changing address back to uint8_t to ensure proper values are being written

+++ MAX30105.cpp
@@ -729,7 +729,7 @@ void MAX30105::bitMask(uint8_t reg, uint8_t mask, uint8_t thing)
 //
 // Low-level I2C Communication
 //
-uint8_t MAX30105::readRegister8(int address, char reg) {
+uint8_t MAX30105::readRegister8(uint8_t address, char reg) {
   uBit.i2c.write(address, &reg, true);
   char temp;
   uBit.i2c.read(address, &temp, 1, false); // Request 1 uint8_t
@@ -739,7 +739,7 @@ uint8_t MAX30105::readRegister8(int address, char reg) {
 
 }
 
-void MAX30105::writeRegister8(int address, char reg, char value) {
+void MAX30105::writeRegister8(uint8_t address, char reg, char value) {
   uBit.i2c.write(address, &reg, 1, true);
   uBit.i2c.write(address, &value, 1, false);
 }
